fixing message

+++ koopey.api/src/main/java/com/koopey/api/model/dto/UserDto.java
@@ -9,7 +9,7 @@ public class UserDto extends BaseDto {
     private String avatar;
     private String currency;
     private String language;
-    private Float average;
+    private Integer average;
     private Integer positive;
     private Integer negative;
     private Number latitude;
